Load external scripts without blocking rendering
External scripts normally force the browser to wait for the JavaScript to be fetched, which may add one or more network roundtrips before the page can be rendered. 

So javaScript blocks DOM construction and thus delays the time to first render. To prevent JavaScript from blocking the parser we use the `async` attribute for external scripts.